started input handling

+++ Update.sh
@@ -3,6 +3,10 @@ cd Source
 GOOS=js GOARCH=wasm go build -o Game.wasm -v *.go
 mv Game.wasm ../Res/
 cd ..
-git add .
-git commit -m "$1" $2
-git push origin main --force
+if [ -f ./Res/Game.wasm ]; then
+  git add .
+  git commit -m "$1" $2
+  git push origin main --force
+else
+  echo "Cannot push WebAssembly does not exist"
+fi
